+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 0.2.0
 ### Changed
 - `IConfig` renamed to `Config`
+- `max_range()` moved to `Config`
 
 
 ## 0.1.0